refactor to make more clear update the tables
